fix: reorganizing the code, for flatter structure and simplicity
