--- src/model_loader/file_loader/mod.rs
@@ -1,6 +1,0 @@
-pub mod file_loader;
-pub mod convert;
-pub mod io;
-pub mod parser;
-pub mod types;
-pub mod tensor_loader;
