+++ src/ops/mod.rs
@@ -1,6 +1,13 @@
-pub mod cpu_features;
+// Core compute kernels (performance-critical, may need SIMD)
 pub mod matmul;
-pub mod rmsnorm;
-pub mod residual_add;
+
+// Activation functions (element-wise transforms)
 pub mod swiglu;
 pub mod softmax;
+
+// Normalization operations
+pub mod rmsnorm;
+
+// Utility functions
+pub mod residual_add;
+pub mod cpu_features;
